Emit ESM for langchain subpath to fix global handler registration
Renamed src/langchain/*.ts to *.mts so tsc outputs .mjs (ESM) alongside .js (CJS).
This fixes the Symbol("lc:configure_hooks") mismatch between CJS and ESM module
instances of @langchain/core, enabling registerConfigureHook/setContextVariable
to work correctly when consumers use ESM.

setup() now automatically registers the handler globally via @langchain/core/context,
so users no longer need to pass getHandler() as callbacks manually.

Also adds package.json exports field for dual CJS/ESM resolution and @langchain/core
as a peerDependency.
