+++ package.json
@@ -4,6 +4,19 @@
   "author": "Orq",
   "main": "./index.js",
   "sideEffects": false,
+  "exports": {
+    ".": {
+      "import": "./index.js",
+      "require": "./index.js",
+      "types": "./index.d.ts"
+    },
+    "./langchain": {
+      "import": "./langchain/index.mjs",
+      "require": "./langchain/index.js",
+      "types": "./langchain/index.d.mts"
+    },
+    "./*": "./*"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/orq-ai/orq-node.git"
@@ -14,7 +27,8 @@
     "prepublishOnly": "npm run build"
   },
   "peerDependencies": {
-    "zod": ">= 3 <4"
+    "zod": ">= 3 <4",
+    "@langchain/core": ">=0.3.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.26.0",
